requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
pypdf
faiss-cpu